bus: mhi: core: Check for SYS_ERROR after resume timeout
It is possible that MHI on device can enter the SYS_ERROR state
and lead to an mhi_pm_resume timeout while it waits for M0 event
to come in. Add an extra check to detect SYS_ERROR after the M0
timeout occurs to avoid sending resume failure to the runtime or
system PM framework.

Change-Id: I5f811302bf688f7d4e2d0b01aefeb8ab92488762
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>