bus: mhi: pci_generic: Handle firmware download error callback
In case the requested file for EDL or SBL image is not found in
the file system at power-up, MHI core driver issues a status
callback. Handle it and use it to power down MHI in the future
since it may need to be powered up later with another file.

Change-Id: Ic5f77dbdfb987ea7b1e227f46225f3b98fb62892
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>