bus: mhi: Update documentation for MHI error callbacks
Design changes prompted redefinition of the FATAL and SYSTEM
ERROR callbacks such that FATAL ERROR should be treated as a
callback arriving from an atomic or interrupt context and taken
as an indication that the device has reset to one of the PBL
execution environments. Whereas, a SYSTEM ERROR (SYS_ERROR)
callback is sleep-capable and allowed to block such that the
controller can pursue its clean-up before core driver handles
further processing. Updated documentation for the same.

Change-Id: Ia46ba0d29d31f51c02683cdab58983da500040b9
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>