bus: mhi: pci_generic: Gracefully shutdown on hibernate
mhi_pci_freeze() called for hibernate entry does not issue an MHI
reset to the device. Hence, the device is unaware that MHI host
is undergoing a power cycle and upon a power on, device will have
a stale MHI context. Ensure a graceful shutdown is done in those
hibernate cases since the link remains accessible as the host
initiates the operation.

Change-Id: Ie094b0e65ad577309f09b26841a6c0731e0f2d8e
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>