bus: mhi: pci_generic: Make pci device wake capable
As of now, the pci_wake_from_d3() API returns a failure and since
there is no check for its return value, the issue goes unnoticed
that the device is unable to become wakeup capable in the event
that any data arrives from the device when PCIe is in D3hot state
via a PME. The PME is ignored by the controller and hence, MHI
host remains unaware of device's intention to trigger a wakeup by
sending data packets to host. Fix this issue by making the device
wakeup capable to ensure any incoming data is seen when MHI is in
a suspended state after PME is received.

Change-Id: I6096b264c9fe1d0d9d8c850de6a9bc8ebae46b18
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>