Fix cross-compile script to output commented header text
Cross-compile script currently outputs a header for installer and
uninstaller scripts to mention that they are auto-generated.
However, it is likely that running those scripts will show some
failures given that bash can interpret the text as commands
instead of comments. Fix that to ensure there are no such issues.

Change-Id: Iddbe3f52683856cff4e01bc4f8d4faf3a9a82f4f
Signed-off-by: Bhaumik Bhatt <quic_bbhatt@quicinc.com>