bus: mhi: pci_generic: Add missing poweroff() PM callback
During hibernation process, once thaw() stage completes, the MHI endpoint
devices will be in M0 state post recovery. After that, the devices will be
powered down so that the system can enter the target sleep state. During
this stage, the PCI core will put the devices in D3hot. But this transition
is not allowed by the MHI spec.

So for fixing this issue, let's add the poweroff() callback that will get
executed before putting the system in target sleep state during
hibernation. This callback will power down the device properly so that it
could be restored during restore() stage or thaw() stage.

We also flush recovery_work to ensure recovery is done before freeze().

Change-Id: I9290141e306d4203ffeefaf0b3774e4d065e1915
Signed-off-by: Qiang Yu <quic_qianyu@quicinc.com>