bus: mhi: core: Add support to pass BEI flag in queue API
Currently BEI flag is set for all TREs based on interrupt
moderation timer value. MHI clients are not allowed to block
event interrupts at runtime. In case, interrupt moderation
is disabled for an event ring, client is allowed to poll
on event posted on event ring by blocking the MSI. If
interrupt moderation is enabled, BEI flag passed in queue
API is overridden, in order to maintain the current
implementation which is tied to interrupt moderation timer
value.

Change-Id: I472139770e1175e2ec97198c56a4eec149f346cc
Signed-off-by: Hemant Kumar <quic_hemantk@quicinc.com>