bus: mhi: core: Add support to read capability id from MISC reg
This API is used to determine the capability supported by device.
The miscellaneous registers are a group of optional registers which
implement various functionalities that are not part of the core MHI
functionality. The host must read the capability ID of the register in
order to determine which functionality is supported. Each capability
will point to the next capability supported by the device.

Change-Id: I616f1f80d4a54aa05e99eb1150752cfd7aaf0595
Signed-off-by: Hemant Kumar <quic_hemantk@quicinc.com>