Update adjfreq for MHI PHC
Add a dummy functionality for phc2sys adjfreq callback.
The older (phc2sys) application requires the adjfreq as
a mandatory callback and that is the reason to add the
dummy callback. The newer (phc2sys) application does not
invoke this API unless explicitly invoked.
To keep it backward compatible with the older application
of phc2sys add the call back.

Change-Id: I66b4adbe9a98f57130b25fc5a1e21efc075d2f70
Signed-off-by: Taniya Das <quic_tdas@quicinc.com>