mhi-host: rpm: Fix regular expressions for fetching devices
lassen_log_gen.sh: Fixed regular expressions for fetching the
correct number of devices attached to the host.

To fetch the number of devices attached, we grep on the mhi interfaces
listed out by the ls command. This fix looks for an exact match
instead of a substring , as was the case earlier.

O/P without fix:
Device0 Device1 Device2 Device3 Device4 Device5

After fix:
Device0.


Change-Id: Ia536f33217847ffdbde24f6ee969518aef6132e4
Signed-off-by: Sushrut Shree Trivedi <quic_sushruts@quicinc.com>