mhi_host: rpm: Convert Lane margining data into json format
lassen_log_parser_script.py: The functionality of the script is
being changed. Earlier, it was used to generate eye diagrams and
output eye width and height into an xml file. Now, it parses the
lane margining data gathered from the device and dumps it in
json format into a text file. This data is finally parsed via
an inbuilt utility in the PCAT tool to generate eye-diagrams.

Sample Output:

[EOMREPORTLIB]{
[EOMREPORTLIB]    "version": "0.0.0",
[EOMREPORTLIB]    "results": [
[EOMREPORTLIB]        {
[EOMREPORTLIB]            "interface": "pcie",
[EOMREPORTLIB]            "instance": 0,
[EOMREPORTLIB]            "time_scale": 1.953,
[EOMREPORTLIB]            "lanes": [
[EOMREPORTLIB]                {
[EOMREPORTLIB]                    "lane_number": 0,
[EOMREPORTLIB]                    "eye": [
[EOMREPORTLIB]                        [
[EOMREPORTLIB]                            0,
[EOMREPORTLIB]                            0,
[EOMREPORTLIB]                            0
[EOMREPORTLIB]                        ],
[EOMREPORTLIB]                        [
[EOMREPORTLIB]                            1,
[EOMREPORTLIB]                            0,
[EOMREPORTLIB]                            0
[EOMREPORTLIB]                        ].

Change-Id: Ia02021bdff9fb5575b4bbfe8bfef3dd20e269ca7