bus: mhi: pci_generic: Increase number of TREs for QDSS for cinder
With existing number of transfer ring elements (TREs) as 64 for QDSS SW
channel, seeing 'no free buffer' error on device. This is resulting in
log drops. To support high traffic from QDSS, increasing the number of
TREs to 1024 for QDSS channel and increasing the corresponding number of
event ring elements from 256 to 2048 for cinder.

Change-Id: Ia4a91dcbb3e0249531752a8229f8fc2ab7eeb1af
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>