bus: mhi: uci: Add support for non-blocking reads and writes
Add support for non-blocking reads to prevent the client driver from
waiting indefinitely for read or write operations to complete.

The QUTS application will open the file descriptor with NONBLOCK read. If
the number of free buffers available is zero while queueing to the device,
the host can return -EAGAIN to the client, allowing the client/application
to requeue the request.

Similarly, for reads, if there is no data available from the device, the
MHI UCI host can return -EAGAIN to the client. This enables the client to
initiate another read operation later.

Change-Id: Ib61609b6407b254763c172cf217f9ddc54061ec7
Signed-off-by: Vivek Pernamitta <vpernami@qti.qualcomm.com>